% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPMbase.r
\name{mapPMbase}
\alias{mapPMbase}
\title{Map PAWMAP base map}
\usage{
mapPMbase(pltArterial = TRUE, ldStreams = FALSE, waterColor = "royalblue2")
}
\arguments{
\item{pltArterial}{Should the arterials be mapped?}

\item{ldStreams}{Should the streams be mapped?  Note: this layer
takes more time to plot}

\item{waterColor}{R color name for rivers and streams}
}
\description{
Map PAWMAP base map
}
