% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPMdata.r
\name{mapPMdata}
\alias{mapPMdata}
\title{Map PAWMAP data}
\usage{
mapPMdata(dfm, vbl, color = NULL, size = c(3, 12))
}
\arguments{
\item{dfm}{Data frame with data to be mapped}

\item{vbl}{The field containing the value to map}

\item{color}{The field containing the color variable.  Typically used to
indicate percent detection. If not specified, the aesthetic is not mapped.}

\item{size}{The size range for the plotted points}
}
\description{
Map PAWMAP data
}

