% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPMdiscr.r
\name{mapPMdiscr}
\alias{mapPMdiscr}
\title{Map PAWMAP categorical data by color}
\usage{
mapPMdiscr(dfm, vbl, legend = NULL, ...)
}
\arguments{
\item{dfm}{Data frame with data to be mapped}

\item{vbl}{The field containing the categorical value to map}

\item{legend}{Text for the map legend}
}
\description{
Map PAWMAP categorical data by color
}
