% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPMrtg.r
\name{mapPMrtg}
\alias{mapPMrtg}
\title{Map PAWMAP data by size w/ quantile indicated by color}
\usage{
mapPMrtg(dfm, vbl, legend = "", ...)
}
\arguments{
\item{dfm}{Data frame with data to be mapped}

\item{vbl}{The field containing the value to map}

\item{legend}{Text for the map legend}
}
\description{
Map PAWMAP data by size w/ quantile indicated by color
}
