% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMapFrame.r
\name{setMapFrame}
\alias{setMapFrame}
\title{Return Axis Limits for a Map}
\usage{
setMapFrame(extent = c("Portland", "West Hills", "Tryon Creek", "NW Streams"))
}
\arguments{
\item{extent}{The extent of the map frame.}
}
\description{
Return Axis Limits for a Map
}
