% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.r
\name{statByStation}
\alias{statByStation}
\title{Calculate summary statistic by station}
\usage{
statByStation(dfm, stat = "mean", station = "loc_code", vbl = "result",
  cens = NULL)
}
\arguments{
\item{dfm}{data frame containing the data}

\item{stat}{statistical function to calculate.  Mean is the default.}

\item{station}{station field name, supplied as a string}

\item{vbl}{field name numeric variable to summarize, supplied as a string}

\item{cens}{optional field name indicating whether the analyte was detected ('=') or not ('<')}
}
\value{
a data frame with summary statistic by station
}
\description{
Calculate summary statistic by station
}
