% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPMbase.r
\name{mapPMbase}
\alias{mapPMbase}
\title{Map PAWMAP base map}
\usage{
mapPMbase(pltArterial = TRUE, ldStreams = FALSE, waterColor = "royalblue2")
}
\arguments{
\item{pltArterial}{Should the arterials shapefile be plotted?}

\item{ldStreams}{Should the streams shapefile be loaded?  Note: this layer
takes more time to plot}

\item{waterColor}{color for rivers and streams}
}
\description{
Map PAWMAP base map
}
