% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPMrtg.r
\name{mapPMrtg}
\alias{mapPMrtg}
\title{Map PAWMAP data by size w/ quantile indicated by color}
\usage{
mapPMrtg(
  dfm,
  vbl,
  legend = "",
  highgood = FALSE,
  nclr = 5,
  by.y = "station",
  ...
)
}
\arguments{
\item{dfm}{Data frame with data to be mapped}

\item{vbl}{The field containing the value to map}

\item{legend}{Text for the map legend}

\item{highgood}{Are high values good (e.g., biotic indices = TRUE) or
bad (e.g., pollutants = FALSE)?  Default is FALSE.}

\item{nclr}{number of colors/breaks in data}

\item{by.y}{field in data which matches with station field in shapefile}
}
\description{
Map PAWMAP data by size w/ quantile indicated by color
}
